@@ -1,12 +1,11 @@
-﻿namespace BlockParty.Blocks.Beam
+﻿namespace BlockParty.Blocks.Beam;
+
+public class BeamBlockSettings
 {
-    public class BeamBlockSettings
-    {
-        /// <summary>
-        /// When BeamBlock is completed it might have a partial window accumulated.<br/>
-        /// This property, when FALSE, will emit that final - potentially partial - window on completition.<br/>
-        /// When TRUE, the final - potentially partial - window will NOT be POSTed to downstream blocks and will be discarded.
-        /// </summary>
-        public bool OmitIncompleteFinalWindow { get; set; }
-    }
+    /// <summary>
+    /// When BeamBlock is completed it might have a partial window accumulated.<br/>
+    /// This property, when FALSE, will emit that final - potentially partial - window on completition.<br/>
+    /// When TRUE, the final - potentially partial - window will NOT be POSTed to downstream blocks and will be discarded.
+    /// </summary>
+    public bool OmitIncompleteFinalWindow { get; set; }
 }
