@@ -1,8 +1,7 @@
-﻿namespace BlockParty.Blocks.SequencePreserving
+﻿namespace BlockParty.Blocks.SequencePreserving;
+
+public class SequencePreservingBlockSettings
 {
-    public class SequencePreservingBlockSettings
-    {
-        public SequenceInitialization SequenceInitialization { get; set; } = SequenceInitialization.FromFirstElement;
-        public OnCompleteBufferedMessageBehavior OnCompleteBufferedMessageBehavior { get; set; } = OnCompleteBufferedMessageBehavior.Discard;
-    }
+    public SequenceInitialization SequenceInitialization { get; set; } = SequenceInitialization.FromFirstElement;
+    public OnCompleteBufferedMessageBehavior OnCompleteBufferedMessageBehavior { get; set; } = OnCompleteBufferedMessageBehavior.Discard;
 }
